Split Main * Move server to its own file
* Move request to its own file
* Move response to its own file
* Move tests to the appropriate files
