+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod server;
+pub mod request;
+pub mod response;
+
+pub use server::start_server;
